% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvalr-package.R
\docType{package}
\name{cvalr}
\alias{cvalr}
\title{cvalr: Credit derivative valuation in R}
\description{
The package contains a framework to valuate credit derivatives (at the moment
portfolio CDS swaps and CDO tranches) for various multivariate models (e.g.
exchangeable and hierarchical Marshall--Olkin / Archimedean / Gaussian
models).
}
\section{Calibration parameter}{

A * calibration parameter* represents a family of multivariate models.  or
each model the method \code{expected_value()} can be used to calculate arbitrary
xpectations on the \emph{default counting process}. The implementation depends on
the model and can involve Monte-Carlo simulation. For specific derivatives,
the methods \verb{expected_*_loss}, e.g. \code{\link[=expected_pcds_loss]{expected_pcds_loss()}} and
\code{\link[=expected_cdo_loss]{expected_cdo_loss()}}, should be used to allow for optimized calculations of
the expectation (in particular for the Gaussian case).
\itemize{
\item At the moment, all models have exponential margins. This is not a
limitation for derivatives which are not path dependent (payoffs only
depend on state values at the payoff date), as the time can be transformed
to account for other margins.
\item Most exchangeable models are implemented with two factors \code{lambda} and
\code{nu}, where the former is the marginal rate and the latter is a dependence
parameter which is able to represent a range between zero and one for all
classical bivariate dependence measures. The methods \code{setAlpha()},
\code{setRho()}, and \code{setTau()} can be used to set the dependence parameter such
that the model has a certain \emph{(lower) tail-correlation coefficient},
\emph{Spearman's Rho}, or \emph{Kendall's Tau}.
}
}

