% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-parameter.R
\docType{class}
\name{ExMarkovParam-class}
\alias{ExMarkovParam-class}
\alias{ExMarkovParam}
\title{Exchangeable Markovian calibration parameter}
\description{
Calibration parameter class for the general exchangeable model with a
Markovian \emph{default counting process}.
}
\details{
The probability of \eqn{j > i} portfolio items being defaulted at time
\eqn{t > s} conditioned on \eqn{i} portfolio items being defaulted at time
\eqn{s} is

\deqn{
  \mathbb{P}(Z_t = j \mid Z_s = i)
    = \delta_{i}^\top \operatorname{e}^{(t-s) Q} \delta_{j} .
}
}
\section{Slots}{

\describe{
\item{\code{qmatrix}}{The \eqn{(d+1) \times (d+1)} Markov-generator matrix of the
default counting process}
}}

\examples{
ExMarkovParam(
 qmatrix = matrix(
   c(-0.07647059, 0, 0, 0.05294118, -0.05, 0, 0.02352941, 0.05, 0),
   nrow = 3, ncol = 3))

}
