% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-parameter.R
\docType{methods}
\name{expected_loss}
\alias{expected_loss}
\alias{expected_loss,ExMarkovParam-method}
\alias{expected_loss,ExMarkovParam}
\title{Expected loss for calibration parameter}
\usage{
expected_loss(object, t, g, ...)

\S4method{expected_loss}{ExMarkovParam}(object, t, g, ...)
}
\arguments{
\item{object}{The calibration parameter object}

\item{t}{Point-in-time}

\item{g}{Payoff function}

\item{...}{Further arguments to \code{g}}
}
\description{
Calculates for a function \eqn{g} and the average default counting process
\eqn{L} the expectation
\deqn{
  \mathbb{E}[g(L_t)] .
}
}
\details{
For a portfolio CDS choose \eqn{g(x) = (1 - R) x} and for a CDO tranche with
attachment points \eqn{l < u} and choose
\eqn{g(x) = min{\{ \max{\{ (1 - R) x - l, 0 \}}, u - l \}}}, where \eqn{R} is
the recovery rate.
}
\examples{
expected_loss(CuadrasAugeExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3,
  function(x) pmin(pmax(0.6 * x - 0.1, 0), 0.2))
expected_loss(AlphaStableExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3,
  function(x) pmin(pmax(0.6 * x - 0.1, 0), 0.2))
expected_loss(PoissonExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3,
  function(x) pmin(pmax(0.6 * x - 0.1, 0), 0.2))
expected_loss(ExponentialExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3,
  function(x) pmin(pmax(0.6 * x - 0.1, 0), 0.2))

}
