% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-parameter.R
\docType{class}
\name{ExtGaussian2FParam-class}
\alias{ExtGaussian2FParam-class}
\alias{ExtGaussian2FParam}
\title{Two-factor extendible Gaussian calibration parameter classes}
\description{
Calibration parameter classes with two parameters for the extendible
Gaussian equi-correlation family with exponential margins.
}
\details{
For all implemented families, the parameter \code{nu} can be replaced by
\emph{Spearman's Rho} \code{rho}, \emph{Kendall' Tau} \code{tau}.
For all implemented families, the possible range for \code{rho} and \code{tau}
is from zero to one (boundaries might not be included) and have a
one-to-one mapping to the model-specific parameter \code{nu}.
The link between the Pearson correlation coefficient and
Spearman's Rho and Kendall's Tau is
\itemize{
\item \eqn{\rho = 2 \sin(\rho_S \cdot \pi / 6)} and
\eqn{\rho_S = 6 / \pi \cdot \arcsin(\rho/2)}
\item \eqn{\rho = \sin(\tau \cdot \pi / 2)} and
\eqn{\tau = 2 / \pi \cdot \arcsin(\rho)}
}
}
\section{Slots}{

\describe{
\item{\code{lambda}}{The marginal rate}

\item{\code{nu}}{Model specific parameter (Pearson correlation)}
}}

\examples{
ExtGaussian2FParam(dim = 2L, lambda = 0.05, rho = 0.4)

}
