% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R
\docType{class}
\name{H2ExCalibrationParam-class}
\alias{H2ExCalibrationParam-class}
\title{Virtual superclass for H2-exchangeable calibration parameters}
\description{
A virtual superclass for all 2-level hierarchically-exchangeable calibration
parameters for multivariate (portfolio) credit models.
}
\section{Slots}{

\describe{
\item{\code{partition}}{Partition of the components (only adjacent grouping allowed)}
}}

