% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R
\docType{class}
\name{H2ExMOParam-class}
\alias{H2ExMOParam-class}
\alias{H2ExMOParam}
\title{H2-Exchangeable Marshall--Olkin calibration parameter}
\description{
Calibration parameter class for the general 2-level
hierarchically-exchangeable model from the Marshall--Olkin class.
}
