% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-H2ExtGaussian3FParam.R
\docType{class}
\name{H2ExtGaussian3FParam-class}
\alias{H2ExtGaussian3FParam-class}
\alias{H2ExtGaussian3FParam}
\alias{simulate_dt,H2ExtGaussian3FParam-method}
\title{Three-factor H2-extendible Gaussian calibration parameter classes}
\usage{
\S4method{simulate_dt}{H2ExtGaussian3FParam}(object, ..., n_sim = 10000)
}
\arguments{
\item{object}{A \linkS4class{CalibrationParam}-object.}

\item{...}{Pass-through parameter.}

\item{n_sim}{Number of samples.}
}
\description{
Calibration parameter classes with three parameters for the 2-level
hierarchically-extendible Gaussian family with exponential margins.
}
\details{
For all implemented families, the parameters \code{nu} can be replaced by
\emph{Spearman's Rho} \code{rho}, \emph{Kendall's Tau} \code{tau}, or the \emph{(lower) tail
dependence coefficient} \code{alpha}.
For all implemented families, the possible range for \code{rho}, \code{tau}, and
\code{alpha} is between zero and one with the restriction that the global
parameter has to be smaller or equal than the corresponding component
parameter. Additionally, we support only that parameters of the same type are
provided, i.e. \code{rho}. The parameters have a one-to-one mapping to \code{nu}.
}
\section{Functions}{
\itemize{
\item \code{simulate_dt,H2ExtGaussian3FParam-method}: simulates the default times \eqn{(\tau_1, \ldots, \tau_d)} and returns a
matrix \code{x} with \code{nrow(x) == n_sim} and \code{ncol(x) == dim(object)} if
\code{dim(object) > 1L} and a vector \code{x} with \code{length(x) == n_sim} otherwise.
}}

\section{Slots}{

\describe{
\item{\code{lambda}}{The marginal rate}

\item{\code{nu}}{Model-specific dependence parameters for the global- and the
component-models.}
}}

\examples{
parm <- H2ExtGaussian3FParam(
  partition = list(1:2, 3:6, 7:8),
  lambda = 8e-2, rho = c(0.2, 0.7))
simulate_dt(parm, n_sim = 5e1)

}
