% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R
\docType{class}
\name{H2ExtMO3FParam-class}
\alias{H2ExtMO3FParam-class}
\alias{CuadrasAugeH2ExtMO3FParam-class}
\alias{CuadrasAugeH2ExtMO3FParam}
\alias{AlphaStableH2ExtMO3FParam-class}
\alias{AlphaStableH2ExtMO3FParam}
\alias{PoissonH2ExtMO3FParam-class}
\alias{PoissonH2ExtMO3FParam}
\alias{ExponentialH2ExtMO3FParam-class}
\alias{ExponentialH2ExtMO3FParam}
\title{Two-factor H2-extendible Marshall--Olkin calibration parameter classes}
\description{
Calibration parameter class with three parameters for the general 2-level
hierarchically-etendiblew model from the Marshall--Olkin class.
}
\details{
For all implemented families, the parameters \code{nu}can be replaced by
\emph{Spearman's Rho} \code{rho}, \emph{Kendall's Tau} \code{tau}, or the \emph{(lower) tail
dependence coefficient} \code{alpha}.
For all implemented families, the possible range for \code{rho}, \code{tau}, and
\code{alpha}, is between zero and one with the restriction that the global
parameter has to be smaller or equal than the corresponding component
parameter. Additionally, we support only that parameters of the same type are
provided, i.e. \code{rho}. The parameters have a one-to-one mapping to \code{nu}.
}
\section{Slots}{

\describe{
\item{\code{lambda}}{The marginal rate}

\item{\code{nu}}{Model-specific dependence parameters for the global- and the
component-models.}
}}

\section{Cuadras-Augé calibration parameter class}{

Corresponds to a Lévy subordinators which are a convex combination of
a pure-killing subordinator and a pure-drift subordinator.
\itemize{
\item \eqn{\psi(x) = \nu + (1 - \nu) x}
\item \eqn{\alpha = \nu}
}
}

\section{Alpha-stable calibration parameter class}{

Corresponds to \eqn{\alpha}-stable subordinators.
\itemize{
\item \eqn{\psi(x) = x^\nu}
\item \eqn{\nu = \log_2(2 - \alpha)} and \eqn{\alpha = 2 - 2^\nu}
}
}

\section{Poisson calibration parameter class}{

Corresponds to Lévy subordinators which are a convex combination of
Poisson subordinators with jump size \code{nu} and pure-drift subordinators.
\itemize{
\item \eqn{\psi(x) = \operatorname{e}^{-\nu}x + (1 - \operatorname{e}^{-x \nu})}
\item \eqn{\nu = -log(1 - sqrt(\alpha))} and \eqn{\alpha = (1 - \operatorname{e}^{-\eta})}
}
}

\section{Exponential calibration parameter class}{

Corresponds to Lévy subordinator which is a convex combination of
Exponential-jump compound Poisson processes with rate \code{nu} and unit-intensity
and a pure-drift subordinators.
\itemize{
\item \eqn{\psi(x) = (1 - 1 / (1 + \nu))x + 1 / (x + \nu)}
\item \eqn{\nu = 0.5 \cdot (-3 + \sqrt{1 + 8 / \alpha})}
and \eqn{\alpha = 2 / (1 + \nu) - 1 / (2 + \nu)}
}
}

