% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-H2ExtMOParam.R
\docType{class}
\name{H2ExtMOParam-class}
\alias{H2ExtMOParam-class}
\alias{H2ExtMOParam}
\title{H2-Extendible Marshall--Olkin calibration parameter}
\description{
\linkS4class{CalibrationParam} for the H2-extendible Marshall-Olkin \emph{(average) default counting
process} model. Extends \linkS4class{H2ExMOParam} and related to \linkS4class{ExMOParam}.
}
