% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{portfolio_cds_spread}
\alias{portfolio_cds_spread}
\title{Portfolio CDS spread}
\usage{
portfolio_cds_spread(
  expected_losses,
  expected_nominals,
  times,
  discount_factors
)
}
\arguments{
\item{expected_losses}{The expected losses of the portfolio (accounting
for recovery) at \code{times}}

\item{expected_nominals}{The expected nominal values of the portfolio
(accounting for expected accrued interest) at \code{times}}

\item{times}{The times of the payment schedule}

\item{discount_factors}{The discount factors.}
}
\description{
Portfolio CDS spread
}
