% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allProbability.R
\docType{methods}
\name{probability_distribution}
\alias{probability_distribution}
\alias{probability_distribution,ExMarkovParam-method}
\alias{probability_distribution,ExMarkovParam}
\alias{probability_distribution,ExtGaussian2FParam-method}
\alias{probability_distribution,ExtGaussian2FParam}
\alias{probability_distribution,FrankExtArch2FParam-method}
\alias{probability_distribution,FrankExtArch2FParam}
\title{Probability vector for calibration parameter}
\usage{
probability_distribution(object, times)

\S4method{probability_distribution}{ExMarkovParam}(object, times)

\S4method{probability_distribution}{ExtGaussian2FParam}(object, times)

\S4method{probability_distribution}{FrankExtArch2FParam}(object, times)
}
\arguments{
\item{object}{The calibration parameter object}

\item{times}{Point-in-time}
}
\description{
Calculates the probability vector  of the average default counting
process \eqn{L}.
}
\examples{
probability_distribution(CuadrasAugeExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3)
probability_distribution(AlphaStableExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3)
probability_distribution(PoissonExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3)
probability_distribution(ExponentialExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3)

probability_distribution(ExtGaussian2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3)

probability_distribution(FrankExtArch2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3)

}
