% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allSimulate-S4.R
\name{simulate_param}
\alias{simulate_param}
\alias{simulate_param,ExMarkovParam-method}
\alias{simulate_param,ExMarkovParam}
\alias{simulate_param,ExMOParam-method}
\alias{simulate_param,ExMOParam}
\alias{simulate_param,ExtGaussian2FParam-method}
\alias{simulate_param,ExtGaussian2FParam}
\alias{simulate_param,FrankExtArch2FParam-method}
\alias{simulate_param,FrankExtArch2FParam}
\title{Simulate the default counting process}
\usage{
simulate_param(object, n, times, ...)

\S4method{simulate_param}{ExMarkovParam}(object, n, times, ...)

\S4method{simulate_param}{ExMOParam}(object, n, times, ...)

\S4method{simulate_param}{ExtGaussian2FParam}(object, n, times, ...)

\S4method{simulate_param}{FrankExtArch2FParam}(object, n, times, ...)
}
\arguments{
\item{object}{Calibration parameter object}

\item{n}{Number of samples}

\item{times}{The times for which the process is supposed to be simulated}

\item{...}{Further parameters}
}
\description{
Simulation of the average default counting process \eqn{L} for a given
time grid.
}
\examples{
simulate_param(ExMarkovParam(), 1e1, seq(0, 5, by = 0.25))

simulate_param(ExMOParam(), 1e1, seq(0, 5, by = 0.25))

simulate_param(ExtGaussian2FParam(dim = 5), 1e1, seq(0, 5, by = 0.25))

simulate_param(FrankExtArch2FParam(dim = 5), 1e1, seq(0, 5, by = 0.25))

}
