% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{CalibrationParam-class}
\alias{CalibrationParam-class}
\title{Virtual class \code{CalibrationParam} for calibration parameters}
\description{
A virtual superclass for all calibration parameters for (portfolio) credit
models.
}
\details{
It is assumed that the expected value of the credit derivative
payment streams can be represented as a function of \emph{expected losses} and
other non-model dependent quantities.
\emph{Expected losses} are the expected values of the average default counting
process \eqn{L} under a transformation \eqn{g} on times \eqn{t_1, \ldots, t_n},
i.e. \eqn{E[g(L_{t_1})], \ldots, E[g(L_{t_n})]}.

A \code{CalibrationParam} object is supposed to be used to abstract the calculation
of these \emph{expected losses} in functions which calculate the fair value of the
option with the asset pricing theorem.
}
\section{Slots}{

\describe{
\item{\code{dim}}{The dimension (no of portfolio items)}
}}

\seealso{
\linkS4class{ExMarkovParam} \linkS4class{ExMOParam} \linkS4class{ExtMOParam}
\linkS4class{CuadrasAugeExtMO2FParam} \linkS4class{AlphaStableExtMO2FParam}
\linkS4class{PoissonExtMO2FParam} \linkS4class{ExponentialExtMO2FParam}
}
