% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R, R/simulate_param-S4.R,
%   R/expected_value-S4.R, R/probability_distribution-S4.R
\docType{class}
\name{CalibrationParam-class}
\alias{CalibrationParam-class}
\alias{simulate_param}
\alias{expected_value}
\alias{expected_value,CalibrationParam-method}
\alias{expected_pcds_loss}
\alias{expected_pcds_loss,CalibrationParam-method}
\alias{expected_cdo_loss}
\alias{expected_cdo_loss,CalibrationParam-method}
\alias{expected_pcds_equation}
\alias{expected_pcds_equation,CalibrationParam-method}
\alias{expected_cdo_equation}
\alias{expected_cdo_equation,CalibrationParam-method}
\alias{probability_distribution}
\title{Virtual super-class for calibration parameters}
\usage{
simulate_param(object, times, ...)

expected_value(object, times, g, ...)

\S4method{expected_value}{CalibrationParam}(
  object,
  times,
  g,
  ...,
  method = c("default", "CalibrationParam"),
  pd_args = NULL
)

expected_pcds_loss(object, times, recovery_rate, ...)

\S4method{expected_pcds_loss}{CalibrationParam}(
  object,
  times,
  recovery_rate,
  ...,
  method = c("default", "CalibrationParam")
)

expected_cdo_loss(object, times, recovery_rate, lower, upper, ...)

\S4method{expected_cdo_loss}{CalibrationParam}(
  object,
  times,
  recovery_rate,
  lower,
  upper,
  ...,
  method = c("default", "CalibrationParam")
)

expected_pcds_equation(
  object,
  times,
  discount_factors,
  recovery_rate,
  coupon,
  upfront,
  ...
)

\S4method{expected_pcds_equation}{CalibrationParam}(
  object,
  times,
  discount_factors,
  recovery_rate,
  coupon,
  upfront,
  ...
)

expected_cdo_equation(
  object,
  times,
  discount_factors,
  recovery_rate,
  lower,
  upper,
  coupon,
  upfront,
  ...
)

\S4method{expected_cdo_equation}{CalibrationParam}(
  object,
  times,
  discount_factors,
  recovery_rate,
  lower,
  upper,
  coupon,
  upfront,
  ...
)

probability_distribution(object, times, ...)
}
\arguments{
\item{object}{The calibration parameter object.}

\item{times}{Point-in-time.}

\item{...}{Pass-through parameter.}

\item{g}{Transformation function.}

\item{method}{Calculation method (either \code{"default"} or the name of the
class whose implementation should be used).}

\item{pd_args}{Parameter for \code{\link[=probability_distribution]{probability_distribution()}}.}

\item{recovery_rate}{The recovery rate of the portfolio CDS/CDO.}

\item{lower}{Lower attachment point of the CDO tranche.}

\item{upper}{Upper attachment point of the CDO tranche.}

\item{discount_factors}{Discount factors for \code{times}}

\item{coupon}{Running coupon der CDO tranche.}

\item{upfront}{Upfront payment der CDO tranche.}
}
\description{
A virtual superclass for all calibration parameters for multivariate
(portfolio) credit models.
}
\details{
It is assumed that the expected value of the credit derivative
payment streams can be represented as a function of \emph{expected losses} and
other non-model dependent quantities.
\emph{Expected losses} are the expected values of the average default counting
process \eqn{L} under a transformation \eqn{g} on times \eqn{t_1, \ldots, t_n},
i.e. \eqn{E[g(L_{t_1})], \ldots, E[g(L_{t_n})]}.

A \code{CalibrationParam} object is supposed to be used to abstract the calculation
of these \emph{expected losses} in functions which calculate the fair value of the
option with the asset pricing theorem.
}
\section{Functions}{
\itemize{
\item \code{simulate_param}: simulates the default counting process \eqn{L} and returns a matrix \code{x} with
\code{nrow(x) == n_sim} and \code{ncol(x) == length(times)} if \code{length(times) > 1L}
and a vector \code{x} with \code{length(x) == n_sim} otherwise.

\item \code{expected_value}: returns the expected value for the average default count process \eqn{L} at
a specific time-point.

\item \code{expected_pcds_loss}: returns the expected portfolio CDS loss for a specific time-point.

\item \code{expected_cdo_loss}: returns the expected CDO loss for a specific time-point.

\item \code{expected_pcds_equation}: returns the expected portfolio CDS fair-value equation.

\item \code{expected_cdo_equation}: returns the expected CDO fair-value equation.

\item \code{probability_distribution}: returns the probability vector for the average default count process \eqn{L}.
}}

\section{Slots}{

\describe{
\item{\code{dim}}{The dimension (number of portfolio items)}
}}

\section{Expected value}{

The \emph{expectated value} of the \emph{average default counting process} \eqn{L}
under transformations can be calculated with \code{\link[=expected_value]{expected_value()}}; i.e. the
value
\deqn{
  \mathbb{E}[g(L_t)] , \quad t \geq 0.
}
For a portfolio CDS choose \eqn{g(x) = (1 - R) x} and for a CDO tranche with
attachment points \eqn{l < u} and choose
\eqn{g(x) = min{\{ \max{\{ (1 - R) x - l, 0 \}}, u - l \}}}, where \eqn{R} is
the recovery rate.
}

\section{Probability distribution}{

The probability vector of the \emph{average default counting process} \eqn{L}
for certain times can be calculated with \code{\link[=probability_distribution]{probability_distribution()}};
i.e. the values
\deqn{
  \mathbb{P}(L_t = k/d) , \quad k \in {\{ 0, \ldots, d \}} , \quad t \geq 0 .
}
}

\examples{
expected_value(CuadrasAugeExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3,
  function(x) pmin(pmax(0.6 * x - 0.1, 0), 0.2))
expected_value(AlphaStableExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3,
  function(x) pmin(pmax(0.6 * x - 0.1, 0), 0.2))
expected_value(PoissonExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3,
  function(x) pmin(pmax(0.6 * x - 0.1, 0), 0.2))
expected_value(ExponentialExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3,
  function(x) pmin(pmax(0.6 * x - 0.1, 0), 0.2))

expected_pcds_loss(CuadrasAugeExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3, 0.4)
expected_pcds_loss(AlphaStableExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3, 0.4)
expected_pcds_loss(PoissonExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3, 0.4)
expected_pcds_loss(ExponentialExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3, 0.4)

expected_cdo_loss(CuadrasAugeExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3, 0.4, 0.1, 0.2)
expected_cdo_loss(AlphaStableExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3, 0.4, 0.1, 0.2)
expected_cdo_loss(PoissonExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3, 0.4, 0.1, 0.2)
expected_cdo_loss(ExponentialExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3, 0.4, 0.1, 0.2)

expected_pcds_equation(
  ExtGaussian2FParam(dim = 75, lambda = 0.05, rho = 0.6),
  times = seq(0.25, 5, by = 0.25), discount_factors = rep(1, 20),
  recovery_rate = 0.4, coupon = 0.08, upfront = 0
)
expected_pcds_equation(
  ExtGaussian2FParam(dim = 75, lambda = 0.05, rho = 0.6),
  times = seq(0.25, 5, by = 0.25), discount_factors = rep(1, 20),
  recovery_rate = rep(0.4, 4), coupon = rep(0.08, 4), upfront = rep(0, 4)
)

expected_cdo_equation(
  ExtGaussian2FParam(dim = 75, lambda = 0.05, rho = 0.6),
  times = seq(0.25, 5, by = 0.25), discount_factors = rep(1, 20),
  recovery_rate = 0.4, lower = c(0, 0.1, 0.2, 0.35),
  upper = c(0.1, 0.2, 0.35, 1), coupon = 0.08, upfront = 0
)

}
\seealso{
\linkS4class{ExMarkovParam} \linkS4class{ExMOParam} \linkS4class{ExtMOParam}
\linkS4class{CuadrasAugeExtMO2FParam} \linkS4class{AlphaStableExtMO2FParam}
\linkS4class{PoissonExtMO2FParam} \linkS4class{ExponentialExtMO2FParam}
}
