% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cdx}
\alias{cdx}
\alias{portfolio_cds_spread}
\alias{upfront_payment}
\alias{upfront_spread}
\title{Portfolio CDS spread and CDO methods}
\usage{
portfolio_cds_spread(expected_losses, times, discount_factors, recovery_rate)

upfront_payment(expected_losses, times, discount_factors, lower, upper, spread)

upfront_spread(expected_losses, times, discount_factors, lower, upper)
}
\arguments{
\item{expected_losses}{The expected losses (not including recovered part)}

\item{times}{The times of the payment schedule}

\item{discount_factors}{The discount factors.}

\item{recovery_rate}{The recovery rate}

\item{lower}{Lower attachement of the CDO tranche}

\item{upper}{Upper attachement of the CDO tranche}

\item{spread}{Constant spread of the CDO tranche}
}
\description{
Portfolio CDS spread and CDO methods
}
