% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cdx}
\alias{cdx}
\alias{portfolio_cds_spread}
\title{Portfolio CDS spread and CDO methods}
\usage{
portfolio_cds_spread(
  expected_default_counts,
  times,
  discount_factors,
  recovery_rate
)
}
\arguments{
\item{expected_default_counts}{The expected default counts of the portfolio
at \code{times}}

\item{times}{The times of the payment schedule}

\item{discount_factors}{The discount factors.}

\item{recovery_rate}{The recovery rate}
}
\description{
Portfolio CDS spread and CDO methods
}
