% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cdx}
\alias{cdx}
\alias{portfolio_cds_coupon}
\alias{portfolio_cds_upfront}
\alias{portfolio_cds_equation}
\alias{cdo_upfront}
\alias{cdo_coupon}
\alias{cdo_equation}
\title{Portfolio CDS spread and CDO methods}
\usage{
portfolio_cds_coupon(expected_losses, times, discount_factors, recovery_rate)

portfolio_cds_upfront(
  expected_losses,
  times,
  discount_factors,
  recovery_rate,
  coupon
)

portfolio_cds_equation(
  expected_losses,
  times,
  discount_factors,
  recovery_rate,
  coupon,
  upfront
)

cdo_upfront(expected_losses, times, discount_factors, lower, upper, coupon)

cdo_coupon(expected_losses, times, discount_factors, lower, upper)

cdo_equation(
  expected_losses,
  times,
  discount_factors,
  lower,
  upper,
  coupon,
  upfront
)
}
\arguments{
\item{expected_losses}{The expected losses (not including recovered part)}

\item{times}{The times of the payment schedule}

\item{discount_factors}{The discount factors.}

\item{recovery_rate}{The recovery rate}

\item{coupon}{The periodically paid coupon of the payment leg}

\item{upfront}{The upfront payment}

\item{lower}{Lower attachment of the CDO tranche}

\item{upper}{Upper attachment of the CDO tranche}
}
\description{
Portfolio CDS spread and CDO methods
}
