% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R, R/allProbability-S4.R
\docType{class}
\name{ExMarkovParam-class}
\alias{ExMarkovParam-class}
\alias{ExMarkovParam}
\alias{probability_distribution,ExMarkovParam-method}
\title{Exchangeable Markovian calibration parameter}
\usage{
\S4method{probability_distribution}{ExMarkovParam}(object, times)
}
\arguments{
\item{object}{The calibration parameter object}

\item{times}{Point-in-time}
}
\description{
Calibration parameter class for the general exchangeable model with a
Markovian \emph{default counting process}.
}
\details{
The probability of \eqn{j > i} portfolio items being defaulted at time
\eqn{t > s} conditioned on \eqn{i} portfolio items being defaulted at time
\eqn{s} is

\deqn{
  \mathbb{P}(Z_t = j \mid Z_s = i)
    = \delta_{i}^\top \operatorname{e}^{(t-s) Q} \delta_{j} .
}
}
\section{Functions}{
\itemize{
\item \code{probability_distribution,ExMarkovParam-method}: returns the probability vector for the average default count process \eqn{L}.
}}

\section{Slots}{

\describe{
\item{\code{qmatrix}}{The \eqn{(d+1) \times (d+1)} Markov-generator matrix of the
default counting process}
}}

\examples{
ExMarkovParam(
 qmatrix = matrix(
   c(-0.07647059, 0, 0, 0.05294118, -0.05, 0, 0.02352941, 0.05, 0),
   nrow = 3, ncol = 3))

probability_distribution(CuadrasAugeExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3)
probability_distribution(AlphaStableExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3)
probability_distribution(PoissonExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3)
probability_distribution(ExponentialExtMO2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3)

}
