% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-ExMarkovParam.R, R/allInitialize-S4.R,
%   R/simulate_param-S4.R, R/probability_distribution-S4.R
\docType{class}
\name{ExMarkovParam-class}
\alias{ExMarkovParam-class}
\alias{ExMarkovParam}
\alias{initialize,ExMarkovParam-method}
\alias{initialize,ExMarkovParam,ANY-method}
\alias{simulate_dt,ExMarkovParam-method}
\alias{probability_distribution,CalibrationParam-method}
\alias{probability_distribution,ExMarkovParam-method}
\title{Exchangeable Markovian calibration parameter}
\usage{
\S4method{initialize}{ExMarkovParam}(.Object, ex_qmatrix)

\S4method{simulate_dt}{ExMarkovParam}(object, ..., method = c("default", "ExMarkovParam"), n_sim = 10000)

\S4method{probability_distribution}{CalibrationParam}(
  object,
  times,
  ...,
  method = c("default", "CalibrationParam"),
  seed = NULL,
  sim_args = NULL
)

\S4method{probability_distribution}{ExMarkovParam}(
  object,
  times,
  ...,
  method = c("default", "ExMarkovParam", "CalibrationParam")
)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{ex_qmatrix}{(Exchangeable) Q-matrix of the default counting process.}

\item{object}{The calibration parameter object.}

\item{...}{arguments to specify properties of the new object, to
      be passed to \code{initialize()}.}

\item{method}{Calculation method (either \code{"default"} or the name of the
class whose implementation should be used).}

\item{n_sim}{Number of samples.}

\item{times}{Point-in-time.}

\item{seed}{Numeric number (if not NULL, is used to set the seed prior to
Monte-Carlo estimation of probability distribution).}

\item{sim_args}{List with pass-through parameters for \code{\link[=simulate_adcp]{simulate_adcp()}}.}
}
\description{
Calibration parameter class for the general exchangeable model with a
Markovian \emph{default counting process}.
}
\details{
The probability of \eqn{j > i} portfolio items being defaulted at time
\eqn{t > s} conditioned on \eqn{i} portfolio items being defaulted at time
\eqn{s} is

\deqn{
  \mathbb{P}(Z_t = j \mid Z_s = i)
    = \delta_{i}^\top \operatorname{e}^{(t-s) Q} \delta_{j} .
}
}
\section{Functions}{
\itemize{
\item \code{initialize,ExMarkovParam-method}: Constructor

\item \code{simulate_dt,ExMarkovParam-method}: simulates the default times \eqn{(\tau_1, \ldots, \tau_d)} and returns a
matrix \code{x} with \code{nrow(x) == n_sim} and \code{ncol(x) == dim(object)} if
\code{dim(object) > 1L} and a vector \code{x} with \code{length(x) == n_sim} otherwise.

\item \code{probability_distribution,CalibrationParam-method}: returns the probability vector for the average default count process \eqn{L}.

\item \code{probability_distribution,ExMarkovParam-method}: returns the probability vector for the average default count process \eqn{L}.
}}

\section{Slots}{

\describe{
\item{\code{qmatrix}}{The \eqn{(d+1) \times (d+1)} Markov generator matrix of the
default counting process.}
}}

\examples{
ExMarkovParam(
 ex_qmatrix = matrix(
   c(-0.07647059, 0, 0, 0.05294118, -0.05, 0, 0.02352941, 0.05, 0),
   nrow = 3L, ncol = 3L))
parm <- ExMarkovParam(
 ex_qmatrix = matrix(
   c(-0.07647059, 0, 0, 0.05294118, -0.05, 0, 0.02352941, 0.05, 0),
   nrow = 3L, ncol = 3L))
simulate_dt(parm, n_sim = 5e1)

probability_distribution(CuadrasAugeExtMO2FParam(
  dim = 50L, lambda = 0.05, rho = 0.4), 0.3)
probability_distribution(AlphaStableExtMO2FParam(
  dim = 50L, lambda = 0.05, rho = 0.4), 0.3)
probability_distribution(PoissonExtMO2FParam(
  dim = 50L, lambda = 0.05, rho = 0.4), 0.3)
probability_distribution(ExponentialExtMO2FParam(
  dim = 50L, lambda = 0.05, rho = 0.4), 0.3)

}
