% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R, R/simulateParam-S4.R
\docType{class}
\name{ExMOParam-class}
\alias{ExMOParam-class}
\alias{ExMOParam}
\alias{simulate_param,ExMOParam-method}
\title{Exchangeable Marshall--Olkin calibration parameter}
\usage{
\S4method{simulate_param}{ExMOParam}(
  object,
  times,
  ...,
  method = c("default", "ExMOParam", "ExMarkovParam"),
  n_sim = 10000
)
}
\arguments{
\item{object}{The calibration parameter object.}

\item{times}{Point-in-time.}

\item{...}{Pass-through parameter.}

\item{method}{Simulation method (either \code{"default"} or the name of the
class whose implementation should be used).}

\item{n_sim}{Number of samples.}
}
\description{
Calibration parameter classfor the general exchangeable model from the
Marshall--Olkin class.
}
\details{
The joint survival function of all portfolio items is assumed to be
\deqn{
  P(\tau > t)
    = \exp{(- a_0 t_{[1]} - \cdots - a_{d-1} t_{[d]})} ,
}
for \eqn{t_{[1]} \geq \cdots \geq t_{[d]}} begin the descendingly ordered
version of \eqn{t} and
\deqn{
  a_{i}
    = \sum_{l=0}^{d-i-1} \binom{d-i-1}{l} \lambda_{l+1} .
}
}
\section{Functions}{
\itemize{
\item \code{simulate_param,ExMOParam-method}: simulates the default counting process \eqn{L} and returns a matrix \code{x} with
\code{nrow(x) == n_sim} and \code{ncol(x) == length(times)} if \code{length(times) > 1L}
and a vector \code{x} with \code{length(x) == n_sim} otherwise.
}}

\section{Slots}{

\describe{
\item{\code{ex_intensities}}{The exchangeable intensities (see details)}
}}

\examples{
ExMOParam(ex_intensities = c(0.02647059, 0.02352941))

simulate_param(ExMOParam(), 1e1, seq(0, 5, by = 0.25))

}
