% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{ExMOParam-class}
\alias{ExMOParam-class}
\alias{ExMOParam}
\title{Exchangeable Marshall--Olkin calibration parameter}
\description{
Calibration parameter classfor the general exchangeable model from the
Marshall--Olkin class.
}
\details{
The joint survival function of all portfolio items is assumed to be
\deqn{
  P(\tau > t)
    = \exp{(- a_0 t_{[1]} - \cdots - a_{d-1} t_{[d]})} ,
}
for \eqn{t_{[1]} \geq \cdots \geq t_{[d]}} begin the descendingly ordered
version of \eqn{t} and
\deqn{
  a_{i}
    = \sum_{l=0}^{d-i-1} \binom{d-i-1}{l} \lambda_{l+1} .
}
}
\section{Slots}{

\describe{
\item{\code{ex_intensities}}{The exchangeable intensities (see details)}
}}

\examples{
ExMOParam(ex_intensities = c(0.02647059, 0.02352941))

}
