% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R, R/allProbability-S4.R,
%   R/allExpectation-S4.R
\docType{class}
\name{ExtArch2FParam-class}
\alias{ExtArch2FParam-class}
\alias{FrankExtArch2FParam-class}
\alias{FrankExtArch2FParam}
\alias{probability_distribution,FrankExtArch2FParam-method}
\alias{expected_pcds_loss,FrankExtArch2FParam-method}
\title{Two-factor extendible Archimedean calibration parameter classes}
\usage{
\S4method{probability_distribution}{FrankExtArch2FParam}(object, times)

\S4method{expected_pcds_loss}{FrankExtArch2FParam}(
  object,
  times,
  recovery_rate,
  method = c("default", "fallback"),
  ...
)
}
\arguments{
\item{object}{The calibration parameter object}

\item{times}{Point-in-time}

\item{recovery_rate}{The recovery rate of the portfolio CDS/CDO.}

\item{...}{Pass-through parameter}
}
\description{
Calibration parameter classes with two parameters for the extendible
Archimedean families with exponential margins.
}
\details{
For all implemented families, the parameter \code{nu} can be replaced by
\emph{Spearman's Rho} \code{rho}, \emph{Kendall's Tau} \code{tau}.
For all implemented families, the possible range for \code{rho} and \code{tau}
is from zero to one (boundaries might not be included) and have a
one-to-one mapping to the model-specific parameter \code{nu}.
}
\section{Functions}{
\itemize{
\item \code{probability_distribution,FrankExtArch2FParam-method}: returns the probability vector for the average default count process \eqn{L}.

\item \code{expected_pcds_loss,FrankExtArch2FParam-method}: returns the expected portfolio CDS loss for a specific time-point.
}}

\section{Slots}{

\describe{
\item{\code{lambda}}{The marginal recovery_rate}

\item{\code{nu}}{Model specific parameter}
}}

\examples{
probability_distribution(FrankExtArch2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3)

expected_pcds_loss(FrankExtArch2FParam(dim = 75, lambda = 0.05, rho = 0.6),
  times = 0.25, recovery_rate = 0.4)
expected_pcds_loss(FrankExtArch2FParam(dim = 75, lambda = 0.05, rho = 0.6),
  times = 0.25, recovery_rate = 0.4, method = "fallback")

}
