% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{ExtArch2FParam-class}
\alias{ExtArch2FParam-class}
\alias{FrankExtArch2FParam-class}
\alias{FrankExtArch2FParam}
\title{Two-factor extendible Archimedean calibration parameter classes}
\description{
Calibration parameter classes with two parameters for the extendible
Archimedean families with exponential margins.
}
\details{
For all implemented families, the parameter \code{nu} can be replaced by
\emph{Spearman's Rho} \code{rho}, \emph{Kendall's Tau} \code{tau}.
For all implemented families, the possible range for \code{rho} and \code{tau}
is from zero to one (boundaries might not be included) and have a
one-to-one mapping to the model-specific parameter \code{nu}.
}
\section{Slots}{

\describe{
\item{\code{lambda}}{The marginal recovery_rate}

\item{\code{nu}}{Model specific parameter}
}}

