% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R, R/allProbability-S4.R,
%   R/allExpectation-S4.R
\docType{class}
\name{ExtGaussian2FParam-class}
\alias{ExtGaussian2FParam-class}
\alias{ExtGaussian2FParam}
\alias{probability_distribution,ExtGaussian2FParam-method}
\alias{expected_pcds_loss,ExtGaussian2FParam-method}
\alias{expected_cdo_loss,ExtGaussian2FParam-method}
\title{Two-factor extendible Gaussian calibration parameter classes}
\usage{
\S4method{probability_distribution}{ExtGaussian2FParam}(object, times)

\S4method{expected_pcds_loss}{ExtGaussian2FParam}(
  object,
  times,
  recovery_rate,
  method = c("default", "fallback"),
  ...
)

\S4method{expected_cdo_loss}{ExtGaussian2FParam}(
  object,
  times,
  recovery_rate,
  lower,
  upper,
  method = c("default", "fallback"),
  ...
)
}
\arguments{
\item{object}{The calibration parameter object}

\item{times}{Point-in-time}

\item{recovery_rate}{The recovery rate of the portfolio CDS/CDO.}

\item{...}{Pass-through parameter}

\item{lower}{Lower attachment point of the CDO tranche.}

\item{upper}{Upper attachment point of the CDO tranche.}
}
\description{
Calibration parameter classes with two parameters for the extendible
Gaussian equi-correlation family with exponential margins.
}
\details{
For all implemented families, the parameter \code{nu} can be replaced by
\emph{Spearman's Rho} \code{rho}, \emph{Kendall' Tau} \code{tau}.
For all implemented families, the possible range for \code{rho} and \code{tau}
is from zero to one (boundaries might not be included) and have a
one-to-one mapping to the model-specific parameter \code{nu}.
The link between the Pearson correlation coefficient and
Spearman's Rho and Kendall's Tau is
\itemize{
\item \eqn{\rho = 2 \sin(\rho_S \cdot \pi / 6)} and
\eqn{\rho_S = 6 / \pi \cdot \arcsin(\rho/2)}
\item \eqn{\rho = \sin(\tau \cdot \pi / 2)} and
\eqn{\tau = 2 / \pi \cdot \arcsin(\rho)}
}
}
\section{Functions}{
\itemize{
\item \code{probability_distribution,ExtGaussian2FParam-method}: returns the probability vector for the average default count process \eqn{L}.

\item \code{expected_pcds_loss,ExtGaussian2FParam-method}: returns the expected portfolio CDS loss for a specific time-point.

\item \code{expected_cdo_loss,ExtGaussian2FParam-method}: returns the expected CDO loss for a specific time-point.
}}

\section{Slots}{

\describe{
\item{\code{lambda}}{The marginal rate}

\item{\code{nu}}{Model specific parameter (Pearson correlation)}
}}

\examples{
ExtGaussian2FParam(dim = 2L, lambda = 0.05, rho = 0.4)

probability_distribution(ExtGaussian2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3)

expected_pcds_loss(ExtGaussian2FParam(dim = 75, lambda = 0.05, rho = 0.6),
  times = 0.25, recovery_rate = 0.4)
expected_pcds_loss(ExtGaussian2FParam(dim = 75, lambda = 0.05, rho = 0.6),
  times = 0.25, recovery_rate = 0.4, method = "fallback")

expected_cdo_loss(ExtGaussian2FParam(dim = 75, lambda = 0.05, rho = 0.6),
  times = 0.25, recovery_rate = 0.4, lower = 0.1, upper = 0.2)
expected_cdo_loss(ExtGaussian2FParam(dim = 75, lambda = 0.05, rho = 0.6),
  times = 0.25, recovery_rate = 0.4, lower = 0.1, upper = 0.2, method = "fallback")

}
