% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R, R/allInitialize-S4.R,
%   R/simulate_param-S4.R, R/expected_value-S4.R,
%   R/probability_distribution-S4.R
\docType{class}
\name{ExtGaussian2FParam-class}
\alias{ExtGaussian2FParam-class}
\alias{ExtGaussian2FParam}
\alias{initialize,ExtGaussian2FParam-method}
\alias{initialize,ExtGaussian2FParam,ANY-method}
\alias{simulate_param,ExtGaussian2FParam-method}
\alias{expected_pcds_loss,ExtGaussian2FParam-method}
\alias{expected_cdo_loss,ExtGaussian2FParam-method}
\alias{probability_distribution,ExtGaussian2FParam-method}
\title{Two-factor extendible Gaussian calibration parameter classes}
\usage{
\S4method{initialize}{ExtGaussian2FParam}(.Object, dim = 2, lambda = 0.1, nu = 0.5, rho = NULL, tau = NULL)

\S4method{simulate_param}{ExtGaussian2FParam}(object, times, ..., n_sim = 10000)

\S4method{expected_pcds_loss}{ExtGaussian2FParam}(
  object,
  times,
  recovery_rate,
  ...,
  method = c("default", "ExtGaussian2FParam", "CalibrationParam")
)

\S4method{expected_cdo_loss}{ExtGaussian2FParam}(
  object,
  times,
  recovery_rate,
  lower,
  upper,
  ...,
  method = c("default", "ExtGaussian2FParam", "CalibrationParam")
)

\S4method{probability_distribution}{ExtGaussian2FParam}(
  object,
  times,
  ...,
  method = c("default", "ExtGaussian2FParam", "CalibrationParam")
)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{dim}{Dimension.}

\item{lambda}{Marginal intensity.}

\item{nu}{Dependence parameter.}

\item{rho}{Spearman's Rho.}

\item{tau}{Kendall's Tau.}

\item{object}{The calibration parameter object.}

\item{times}{Point-in-time.}

\item{...}{arguments to specify properties of the new object, to
      be passed to \code{initialize()}.}

\item{n_sim}{Number of samples.}

\item{recovery_rate}{The recovery rate of the portfolio CDS/CDO.}

\item{method}{Calculation method (either \code{"default"} or the name of the
class whose implementation should be used).}

\item{lower}{Lower attachment point of the CDO tranche.}

\item{upper}{Upper attachment point of the CDO tranche.}
}
\description{
Calibration parameter classes with two parameters for the extendible
Gaussian equi-correlation family with exponential margins.
}
\details{
For all implemented families, the parameter \code{nu} can be replaced by
\emph{Spearman's Rho} \code{rho}, \emph{Kendall' Tau} \code{tau}.
For all implemented families, the possible range for \code{rho} and \code{tau}
is from zero to one (boundaries might not be included) and have a
one-to-one mapping to the model-specific parameter \code{nu}.
The link between the Pearson correlation coefficient and
Spearman's Rho and Kendall's Tau is
\itemize{
\item \eqn{\rho = 2 \sin(\rho_S \cdot \pi / 6)} and
\eqn{\rho_S = 6 / \pi \cdot \arcsin(\rho/2)}
\item \eqn{\rho = \sin(\tau \cdot \pi / 2)} and
\eqn{\tau = 2 / \pi \cdot \arcsin(\rho)}
}
}
\section{Functions}{
\itemize{
\item \code{initialize,ExtGaussian2FParam-method}: Constructor

\item \code{simulate_param,ExtGaussian2FParam-method}: simulates the default counting process \eqn{L} and returns a matrix \code{x} with
\code{nrow(x) == n_sim} and \code{ncol(x) == length(times)} if \code{length(times) > 1L}
and a vector \code{x} with \code{length(x) == n_sim} otherwise.

\item \code{expected_pcds_loss,ExtGaussian2FParam-method}: returns the expected portfolio CDS loss for a specific time-point.

\item \code{expected_cdo_loss,ExtGaussian2FParam-method}: returns the expected CDO loss for a specific time-point.

\item \code{probability_distribution,ExtGaussian2FParam-method}: returns the probability vector for the average default count process \eqn{L}.
}}

\section{Slots}{

\describe{
\item{\code{lambda}}{The marginal rate}

\item{\code{nu}}{Model specific parameter (Pearson correlation)}
}}

\examples{
ExtGaussian2FParam(dim = 2L, lambda = 0.05, rho = 0.4)
simulate_param(ExtGaussian2FParam(dim = 5), 1e1, seq(0, 5, by = 0.25), n_sim = 5e1)

expected_pcds_loss(ExtGaussian2FParam(dim = 75, lambda = 0.05, rho = 0.6),
  times = 0.25, recovery_rate = 0.4)
expected_pcds_loss(ExtGaussian2FParam(dim = 75, lambda = 0.05, rho = 0.6),
  times = 0.25, recovery_rate = 0.4, method = "CalibrationParam")

expected_cdo_loss(ExtGaussian2FParam(dim = 75, lambda = 0.05, rho = 0.6),
  times = 0.25, recovery_rate = 0.4, lower = 0.1, upper = 0.2)
expected_cdo_loss(ExtGaussian2FParam(dim = 75, lambda = 0.05, rho = 0.6),
  times = 0.25, recovery_rate = 0.4, lower = 0.1, upper = 0.2, method = "CalibrationParam")

probability_distribution(ExtGaussian2FParam(
  dim = 50, lambda = 0.05, rho = 0.4), 0.3)

}
