% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{ExtMO2FParam-class}
\alias{ExtMO2FParam-class}
\alias{CuadrasAugeExtMO2FParam-class}
\alias{CuadrasAugeExtMO2FParam}
\alias{AlphaStableExtMO2FParam-class}
\alias{AlphaStableExtMO2FParam}
\alias{PoissonExtMO2FParam-class}
\alias{PoissonExtMO2FParam}
\alias{ExponentialExtMO2FParam-class}
\alias{ExponentialExtMO2FParam}
\title{Two-factor extendible Marshall--Olkin calibration parameter classes}
\description{
Calibration parameter classes with two parameters for the extendible
Marshall--Olkin family.
}
\details{
For all implemented families, the parameter \code{nu} can be replaced by
\emph{Spearman's Rho} \code{rho}, \emph{Kendall' Tau} \code{tau} or the
\emph{(lower) tail dependen coefficient} \code{alpha}.
For all implemented families, the possible range for \code{rho}, \code{tau}, and
\code{alpha} is from zero to one (boundaries might not be included) and have a
one-to-one mapping to the model-specific parameter \code{nu}.
The link between lower tail dependence coefficient \eqn{\alpha} and
Spearman's Rho and Kendall's Tau is
\itemize{
\item \eqn{\alpha = 4 \rho / (3 + \rho)} and \eqn{\rho = 3 \alpha / (4 - \alpha)}
\item \eqn{\alpha = 2 \tau / (1 + \tau)} and \eqn{\tau = \alpha / (2 - \alpha)}
}
}
\section{Slots}{

\describe{
\item{\code{lambda}}{The marginal rate}

\item{\code{nu}}{Model specific dependence parameter}
}}

\section{Cuadras-Augé calibration parameter class}{

Corresponds to a Lévy subordinator which is a convex combination of
a pure-killing subordinator and a pure-drift subordinator.
\itemize{
\item \eqn{\psi(x) = \nu + (1 - \nu) x}
\item \eqn{\alpha = \nu}
}
}

\section{Alpha-stable calibration parameter class}{

Corresponds to an \eqn{\alpha}-stable subordinator.
\itemize{
\item \eqn{\psi(x) = x^\nu}
\item \eqn{\nu = \log_2(2 - \alpha)} and \eqn{\alpha = 2 - 2^\nu}
}
}

\section{Poisson calibration parameter class}{

Corresponds to a Lévy subrodinator which is a convex combination of a
Poisson subordinator with jump size \code{nu} and a pure-drift subordinator.
\itemize{
\item \eqn{\psi(x) = \operatorname{e}^{-\nu}x + (1 - \operatorname{e}^{-x \nu})}
\item \eqn{\nu = -log(1 - sqrt(\alpha))} and \eqn{\alpha = (1 - \operatorname{e}^{-\eta})}
}
}

\section{Exponential calibration parameter class}{

Corresponds to a Lévy subordinator which is a convex combination of an
Exponential-jump compound Poisson process with rate \code{nu} and unit-intensity
and a pure-drift subordinator.
\itemize{
\item \eqn{\psi(x) = (1 - 1 / (1 + \nu))x + 1 / (x + \nu)}
\item \eqn{\nu = 0.5 \cdot (-3 + \sqrt{1 + 8 / \alpha})}
and \eqn{\alpha = 2 / (1 + \nu) - 1 / (2 + \nu)}
}
}

\examples{
CuadrasAugeExtMO2FParam(dim = 2L, lambda = 0.05, rho = 0.4)

AlphaStableExtMO2FParam(dim = 2L, lambda = 0.05, rho = 0.4)

PoissonExtMO2FParam(dim = 2L, lambda = 0.05, rho = 0.4)

ExponentialExtMO2FParam(dim = 2L, lambda = 0.05, rho = 0.4)

}
