% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R, R/allInitialize-S4.R
\docType{class}
\name{ExtMOParam-class}
\alias{ExtMOParam-class}
\alias{ExtMOParam}
\alias{initialize,ExtMOParam-method}
\alias{initialize,ExtMOParam,ANY-method}
\title{Extendible Marshall--Olkin calibration parameter}
\usage{
\S4method{initialize}{ExtMOParam}(.Object, dim, bf)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{dim}{Dimension.}

\item{bf}{Bernstein function.}
}
\description{
Calibration parameter class for the general extendible model from the
Marshall-Olkin class.
}
\details{
The joint survival function of all portfolio items is assumed to be
\deqn{
  P(\tau > t)
    = \exp{(- a_{0} t_{[1]} - \cdots - a_{d-1} t_{[d]})} ,
}
for \eqn{t_{[1]} \geq \cdots \geq t_{[d]}} begin the descendingly ordered
version of \eqn{t} and
\deqn{
  a_{i}
    = \psi{(i+1)} - \psi{(i)} .
}
}
\section{Functions}{
\itemize{
\item \code{initialize,ExtMOParam-method}: Constructor
}}

\section{Slots}{

\describe{
\item{\code{bf}}{Bernstein function (see details)}
}}

\examples{
ExtMOParam(
  dim = 2,
  bf = rmo::ScaledBernsteinFunction(
    scale = 0.05,
    original = rmo::SumOfBernsteinFunctions(
      first = rmo::ConstantBernsteinFunction(constant = 0.4),
      second = rmo::LinearBernsteinFunction(scale = 1 - 0.4))
    ))
}
