% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R
\docType{class}
\name{H2ExMarkovParam-class}
\alias{H2ExMarkovParam-class}
\alias{H2ExMarkovParam}
\title{H2-exchangeable Markovian calibration parameter}
\description{
Calibration parameter class for the general 2-level
hierarchically-exchangeable model with Markovian \emph{default counting processes}
in the exchangeable sub-components.
}
\details{
We assume that \eqn{\tau} has the stochastic representation to be the
component-wise minimum of a global exchangeable Markovian-vector
\eqn{\tau^{(0)}} and a vector \eqn{(\tau^{(1)}, \ldots, \tau^{(J)})} with
independent exchangeable Markovian-vector sub-vectors \eqn{\tau^{(j)}}.
}
\section{Slots}{

\describe{
\item{\code{models}}{A list with the global and component models (of the type
\code{ExMarkovParam-class})}

\item{\code{fraction}}{The proportion associated with the global model, see details.}
}}

