% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass-S4.R, R/simulate_param-S4.R
\docType{class}
\name{H2ExMarkovParam-class}
\alias{H2ExMarkovParam-class}
\alias{H2ExMarkovParam}
\alias{simulate_dt,H2ExMarkovParam-method}
\title{H2-exchangeable Markovian calibration parameter}
\usage{
\S4method{simulate_dt}{H2ExMarkovParam}(object, ...)
}
\arguments{
\item{object}{The calibration parameter object.}

\item{...}{Pass-through parameter.}
}
\description{
Calibration parameter class for the general 2-level
hierarchically-exchangeable model with Markovian \emph{default counting processes}
in the exchangeable sub-components.
}
\details{
We assume that \eqn{\tau} has the stochastic representation to be the
component-wise minimum of a global exchangeable Markovian-vector
\eqn{\tau^{(0)}} and a vector \eqn{(\tau^{(1)}, \ldots, \tau^{(J)})} with
independent exchangeable Markovian-vector sub-vectors \eqn{\tau^{(j)}}.
}
\section{Functions}{
\itemize{
\item \code{simulate_dt,H2ExMarkovParam-method}: simulates the default times \eqn{(\tau_1, \ldots, \tau_d)} and returns a
matrix \code{x} with \code{nrow(x) == n_sim} and \code{ncol(x) == dim(object)} if
\code{dim(object) > 1L} and a vector \code{x} with \code{length(x) == n_sim} otherwise.
}}

\section{Slots}{

\describe{
\item{\code{models}}{A list with the global and component models (of the type
\code{ExMarkovParam-class})}

\item{\code{fraction}}{The proportion associated with the global model, see details.}
}}

\examples{
parm <- ExponentialH2ExtMO3FParam(
  partition = list(1:2, 3:6, 7:8), fraction = 0.4,
  lambda = 8e-2, rho = c(0.2, 0.7))
simulate_dt(parm, n_sim = 5e1)
}
