% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-H2ExMOParam.R
\docType{class}
\name{H2ExMOParam-class}
\alias{H2ExMOParam-class}
\alias{H2ExMOParam}
\title{H2-Exchangeable Marshall--Olkin calibration parameter}
\description{
\linkS4class{CalibrationParam} for the H2-exchangeable Marshall-Olkin \emph{(average) default counting
process} model. Extends \linkS4class{H2ExMarkovParam} and related to \linkS4class{ExMOParam}.
}
