% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-H2ExtGaussian3FParam.R,
%   R/s4-H2ExtArch3FParam.R
\docType{class}
\name{H2ExtGaussian3FParam-class}
\alias{H2ExtGaussian3FParam-class}
\alias{H2ExtGaussian3FParam}
\alias{initialize,H2ExtGaussian3FParam-method}
\alias{initialize,H2ExtGaussian3FParam,ANY-method}
\alias{simulate_dt,H2ExtGaussian3FParam-method}
\alias{show,H2ExtGaussian3FParam-method}
\alias{show,H2ExtArch3FParam-method}
\title{Three-factor H2-extendible Gaussian calibration parameter}
\usage{
\S4method{initialize}{H2ExtGaussian3FParam}(
  .Object,
  composition = c(2L, 3L),
  lambda = 0.1,
  nu = c(0.2, 0.3),
  rho = NULL,
  tau = NULL
)

\S4method{simulate_dt}{H2ExtGaussian3FParam}(object, ..., n_sim = 10000)

\S4method{show}{H2ExtGaussian3FParam}(object)

\S4method{show}{H2ExtArch3FParam}(object)
}
\arguments{
\item{.Object}{ An object:  see the \dQuote{Initialize Methods} section.}

\item{composition}{An integerish vector with the composition.}

\item{lambda}{Marginal intensity.}

\item{nu}{(Internal) \emph{Outer} and \emph{inner} bivariate dependence parameter.}

\item{rho}{\emph{Outer} and \emph{inner} bivariate Spearman's Rho.}

\item{tau}{\emph{Outer} and \emph{inner} bivariate Kendall's Tau.}

\item{object}{A \linkS4class{CalibrationParam}-object.}

\item{...}{arguments to specify properties of the new object, to
      be passed to \code{initialize()}.}

\item{n_sim}{Number of samples.}
}
\description{
\linkS4class{CalibrationParam} for the H2-extendible Gaussian copula with Exponential margin model for
the \emph{(average) default counting process}  with 3 parameter. Extends \linkS4class{H2ExCalibrationParam}
and related to \linkS4class{ExtGaussian2FParam}.
}
\details{
The model is defined by the assumption that the \emph{multivariate default times} \eqn{\tau = (\tau_1,
\ldots, \tau_d)} are from a H2-extendible Gaussian copula model with Exponential margins.
The model is specified by three parameters (in addition to the composition): The \emph{marginal rate}
\code{lambda} and the (internal) \emph{outer} and \emph{inner dependency parameters} \code{nu} (Pearson correlation).
The dependence parameter \code{nu} should not be set by the user; instead they should provide either
\code{rho} (\emph{Spearman's Rho}) or \code{tau} (\emph{Kendall's Tau}).
The parameters \code{rho} or \code{tau} should be between zero and one, of length 2, and non-decreasing;
the first value represents the \emph{outer dependence} between components of different partition
elements and the second value represents the \emph{inner depenence} between components of the same
partition element.
The link between Spearman's Rho or Kendall's Tau and the
internal dependence parameter (Pearson correlation) is
\itemize{
\item \eqn{\rho = 2 \sin(\rho_S \cdot \pi / 6)} and
\eqn{\rho_S = 6 / \pi \cdot \arcsin(\rho/2)}
\item \eqn{\rho = \sin(\tau \cdot \pi / 2)} and
\eqn{\tau = 2 / \pi \cdot \arcsin(\rho)}
}

For details on the underlying extendible model, see \linkS4class{ExtGaussian2FParam}.
}
\section{Functions}{
\itemize{
\item \code{initialize,H2ExtGaussian3FParam-method}: Constructor

\item \code{simulate_dt,H2ExtGaussian3FParam-method}: simulates the vector of \emph{default times} and returns a matrix \code{x} with
\code{dim(x) == c(n_sim, getDimension(object))}.

\item \code{show,H2ExtGaussian3FParam-method}: Display the object.

\item \code{show,H2ExtArch3FParam-method}: Display the object.
}}

\section{Slots}{

\describe{
\item{\code{lambda}}{A non-negative number for the marginal rate.}

\item{\code{nu}}{A numeric vector of length 2 for the model specific dependence parameters (global and
component specific; range depends on specific model). Use \code{rho} or \code{tau} in the constructor to
set dependence parameter.}
}}

\section{Simulation}{

The default times are sampled in a two-stage procedure: First a sample is drawn from the Gaussian
copula whose correlation matrix reflect the inner- and outer-dependency parameters, i.e.
\eqn{\rho_{i j} = \nu_1} if \eqn{i,j} are from different elements of the partition and
\eqn{\rho_{i j} = \nu_2} if \eqn{i j} are from the same element of the partition; then the
results are transformed using \code{\link[stats:Exponential]{stats::qexp()}}.
}

\examples{
H2ExtGaussian3FParam()
H2ExtGaussian3FParam(composition = c(2L, 4L, 2L), lambda = 8e-2, rho = c(3e-1, 5e-1))
H2ExtGaussian3FParam(composition = c(2L, 4L, 2L), lambda = 8e-2, tau = c(3e-1, 5e-1))
parm <- H2ExtGaussian3FParam(composition = c(2L, 4L, 2L), lambda = 8e-2, rho = c(2e-1, 7e-1))
simulate_dt(parm, n_sim = 5L)

}
