% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4-H2ExtMOParam.R
\docType{class}
\name{H2ExtMOParam-class}
\alias{H2ExtMOParam-class}
\alias{H2ExtMOParam}
\title{H2-Extendible Marshall--Olkin calibration parameter}
\description{
Calibration parameter class for the general 2-level
hierarchically-etendiblew model from the Marshall--Olkin class.
}
